TOOL-1099: add users to both Chef orgs
When we used Delivery, we uploaded the user's key to Delivery.  but we
no longer use Delivery. it'd still be nice to take their key and set it
up in the Chef data bag (outhouse) and mchx_auth cookbook (inhouse)
for autobot and general machine access, though maybe that should be a
separate tool.